setup play mode for multi page
