+++ apps/mocksi-lite/content/ListPopup/index.tsx
@@ -40,6 +40,8 @@ const ListPopup = ({ email, onChat, onClose, onLogout }: ListPopupProps) => {
 		dispatch({ event: AppEvent.ENTER_SETTINGS });
 	};
 
+	console.log('rendering list popup');
+
 	return (
 		<Popup
 			email={email}
