+++ apps/mocksi-lite/content/utils/observeUrlChange.ts
@@ -1,0 +1,17 @@
+export const observeUrlChange = (onChange: () => void) => {
+	let oldHref = document.location.href;
+	const body = document.querySelector("body");
+
+	if (!body) {
+		console.error("body not found");
+		return;
+	}
+
+	const observer = new MutationObserver((mutations) => {
+		if (oldHref !== document.location.href) {
+			oldHref = document.location.href;
+			onChange();
+		}
+	});
+	observer.observe(body, { childList: true, subtree: true });
+};
